requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-CKEditor==0.5.1
email-validator==2.1.0.post1
Pillow==10.1.0
python-slugify==8.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.23
alembic==1.12.1
WTForms==3.1.1
requests==2.31.0 